add relative and absolute intensity for some peak class and fix a null
Previous intensity in peak related class will be normalized to 1000 when
filtering small peaks. 
Null pointer bug: when a spectrum contains no peaks, the program would
return a null array of peaks. This cause the setting relative intensity
to crash. In this commit, the null is replaced by a empty array.